requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
diff-match-patch==20200713
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-filter==22.1
django-import-export==3.0.1
django-phonenumber-field==7.0.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
drf-excel==2.2.0
drf-pdf==0.2.0
drf-spectacular==0.25.1
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
MarkupPy==1.14
multidict==6.0.4
numpy==1.23.4
oauthlib==3.2.2
odfpy==1.4.1
openai==0.26.1
openpyxl==3.0.10
pandas==1.5.1
phonenumbers==8.13.3
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
